feat(windows): implement virtual limits, auto-cancel, and robust CAN reception
- Implemented a millis()-based virtual position tracker (0-5000ms) in ACT Node to prevent window motors from indefinitely stalling in AUTO mode.
- Mapped Window Switch (WBP) ADC states correctly through the BCM Python Gateway.
- Added smooth auto-cancel logic to the ACT Node (pressing the opposite direction cancels an active AUTO roll).
- Fixed CAN bus dropping messages during concurrent Hazard/Window use by changing the ACT Node receive loop to while() and removing blocking Serial.print statements.

+++ wbp/WBP_Node/src/window_swich.cpp
@@ -1,6 +1,6 @@
 #include "window_swich.h"
 
-windowState window_swich(int adc_value){
+windowState window_swich(uint16_t adc_value){
     if (adc_value <B_UP_AUTO)
     {
        return windowState::WINDOW_UP_AUTO;
